+++ src/index.js
@@ -74,7 +74,12 @@ const filesProto = {
   getDownloadLinkById: files.getDownloadLinkById,
   getDownloadLink: files.getDownloadLinkByPath, // DEPRECATED, should be removed very soon
   getDownloadLinkByPath: files.getDownloadLinkByPath,
-  getArchiveLink: files.getArchiveLink,
+  getArchiveLink: function (...args) {
+    warn('getArchiveLink is deprecated, use cozy.files.getArchiveLinkByPaths instead.')
+    return files.getArchiveLink(...args)
+  },
+  getArchiveLinkByPaths: files.getArchiveLinkByPaths,
+  getArchiveLinkByIds: files.getArchiveLinkByIds,
   getFilePath: files.getFilePath,
   getCollectionShareLink: files.getCollectionShareLink,
   listTrash: files.listTrash,